move files around, fmt, cleanup
