+++ metricbeat/scripts/mage/docs_collector.go
@@ -126,9 +126,8 @@ func testIfDocsInDir(moduleDir string) bool {
 // compile and run the seprate go script to generate a list of default metricsets.
 // This is done so a compile-time issue in metricbeat doesn't break the docs build
 func getDefaultMetricsets() (map[string][]string, error) {
-
 	runpaths := []string{
-		mage.OSSBeatDir("scripts/msetlists/main.go"),
+		mage.OSSBeatDir("scripts/msetlists/cmd/main.go"),
 		mage.XPackBeatDir("scripts/msetlists/main.go"),
 	}
