+++ metricbeat/scripts/msetlists/msetlists.go
@@ -15,26 +15,27 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package main
+package msetlists
 
 import (
-	"encoding/json"
-	"fmt"
-	"os"
+	"strings"
 
-	_ "github.com/elastic/beats/metricbeat/include"
-	"github.com/elastic/beats/metricbeat/scripts/msetlists/gather"
+	"github.com/elastic/beats/metricbeat/mb"
 )
 
-func main() {
-
-	msList := gather.DefaultMetricsets()
-
-	raw, err := json.MarshalIndent(msList, "", "  ")
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "error Marshalling json: %s\n", err)
-		os.Exit(1)
+// DefaultMetricsets returns a JSON array of all registered default metricsets
+// It depends upon the calling library to actually import or register the metricsets.
+func DefaultMetricsets() map[string][]string {
+	// List all registered modules and metricsets.
+	var defaultMap = make(map[string][]string)
+	for _, mod := range mb.Registry.Modules() {
+		metricSets, err := mb.Registry.DefaultMetricSets(mod)
+		if err != nil && !strings.Contains(err.Error(), "no default metricset for") {
+			continue
+		}
+		defaultMap[mod] = metricSets
 	}
 
-	fmt.Printf("%s\n", string(raw))
+	return defaultMap
+
 }
