+++ x-pack/metricbeat/scripts/msetlists/main.go
@@ -1,3 +1,7 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
+
 package main
 
 import (
@@ -5,14 +9,20 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/elastic/beats/metricbeat/scripts/msetlists/gather"
+	"github.com/elastic/beats/metricbeat/scripts/msetlists"
 
+	"github.com/elastic/beats/libbeat/paths"
+	"github.com/elastic/beats/metricbeat/mb"
 	_ "github.com/elastic/beats/x-pack/metricbeat/include"
+	xpackmb "github.com/elastic/beats/x-pack/metricbeat/mb"
 )
 
 func main() {
+	path := paths.Resolve(paths.Home, "../x-pack/metricbeat/module")
+	lm := xpackmb.NewLightModulesSource(path)
+	mb.Registry.SetSecondarySource(lm)
 
-	msList := gather.DefaultMetricsets()
+	msList := msetlists.DefaultMetricsets()
 
 	raw, err := json.MarshalIndent(msList, "", "  ")
 	if err != nil {
